Algorithms/CryptographicAlgos/HashAlgos/requirements.txt
mmh3
xxhash
blake3
pycryptodome